example/requirements.txt
pytest==5.4.3
selenium==3.14.0
allure-pytest==2.8.18
allure-python-commons==2.8.18
attrs==20.1.0
py==1.8.1
